CLI: Better algorithm to explore resources.
The crawler does not apply a depth-first algorithm. Instead, it applies
a breadth-first algorithm. This is not bad actually, but imagine we have
a menu in our website, thus we will explore the same category entirely
before exploring another one. This is not optimal to detect different
errors on different pages quickly.

So here is what we do.

First, we have the crawler. It fetches a lot of URLs. Each URL is pushed
in a specific “bucket”. Buckets are computed based on the first part of
the pathname of the URL. So for instance, the following URLs:

    /a/b
    /a/b/c
    /a/b/d
    /a/b/e/f
    /m/n
    /m/n/o
    /m/p/q
    /x/y
    /x/y/z

will be dispatched like this:

    [a] => [/a/b, /a/b/c, /a/b/d, /a/b/e/f]
    [m] => [/m/n, /m/n/o, /m/p/q]
    [x] => [x/y, /x/y/z]

Second, each bucket is a queue with 1 concurrency. They run in parallel.
When a URL in a bucket is consumed by the queue, it is pushed in the
test queue.

Third and finally, the test queue. Its concurrency is 4 by default, but
it can be controlled by the `-p`/`--concurrency` option. The test queue
consumed URLs and start a test (pa11y, PhantomJS, reports etc.).

The maximum URLs counter is managed by the test queue. When the counter
reaches its end, then the test queue is killed, in addition to all the
bucket queues.

With this new algorithm, we are able to better explore the web
application.