CLI: Fix progression max. number.
While using the pipe to receive URLs to compute, it is possible to see:
`1/128` even if the total of URLs is 7. Why? Because the
`program.maximumUrls` variable was updated after having read the whole
pipe. Remind that the testing starts as soon as possible, so it can
start before reaching the end of the pipe.

Now, this variable is updated every time to provide a more accurate
progression result to the user.