Dashboard: Remove async data, and dynamic URL.
This patch removes asynchronous data. It means that the index of the
dashboard no longer send a `GET` request to retrieve all
statistics. This sets a common basis for the dashboard as a server, or
as static files.

Also, this patch introduces dynamic URL for reading a report. Instead of
expecting `/report/<id>/<file>`, we now expect `/<root>/<id></file>`,
where `<root>` is defined by the `--root` option. Thus:

      $ ./a11ym-dashboard --root a/b/c/

will define the `<root>` as `c`, so expecting URL of the form:

     /c/<id>/<file>

This also sets a common basis for the dashboard as a server, or as
static files: URL will be the same in both cases.

Finally, all statistics data structure has been simplified to save some bytes.